[script] update waybar only when waybar is running

+++ scripts/mic-mute-toggle.sh
@@ -2,11 +2,6 @@
 
 pulsemixer --id "$(pulsemixer --list-sources | grep 'Name: USB PnP Audio Device Mono' | sed -r 's/^.*ID: ([a-zA-Z0-9\-]+).*$/\1/')" --toggle-mute
 
-case "$XDG_CURRENT_DESKTOP" in
-  [Hh]yprland)
-    pkill -SIGRTMIN+8 waybar;;
-  sway)
-    ;;
-  *)
-    ;;
-esac
+if pgrep -x waybar &>/dev/null; then
+  pkill -SIGRTMIN+8 waybar
+fi
